backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # ── Existing (auth scaffold — kept for paid version) ──────────────────
    "fastapi[standard]>=0.128.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.0.0",
    "motor>=3.7.1",
    "pycryptodome>=3.20.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
    # ── Platform (open source v1) ─────────────────────────────────────────
    "celery[redis]>=5.4.0",
    "aioredis>=2.0.1",
    "python-magic>=0.4.27",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "tensorflow-cpu>=2.16.1",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "nbformat>=5.10.0",
    "gymnasium>=0.29.0",
]
